// apps/api-server/src/core/validation/schemas/auth.schema.ts
import { z } from 'zod';

export const loginSchema = z.object({
  body: z.object({
    email: z.string().email('Invalid email address'),
    password: z.string().min(6, 'Password must be at least 6 characters'),
  }),
});

export const googleLoginSchema = z.object({
  body: z.object({
    googleId: z.string(),
    email: z.string().email('Invalid email address'),
  }),
});